# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a megrez backend server.",
        "title": "Megrez backend API",
        "termsOfService": "http://megrez.run",
        "contact": {
            "name": "Megrez",
            "url": "http://megrez.run",
            "email": "alkaidchen@qq.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/admin/article": {
            "post": {
                "description": "create article",
                "consumes": [
                    "application/json"
                ],
                "summary": "create article",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "body",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateArticleForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/errmsg.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Article"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/admin/article/{id}": {
            "get": {
                "description": "get article by article id",
                "consumes": [
                    "application/json"
                ],
                "summary": "get article by article id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "article id",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/errmsg.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ArticleDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "description": "update article",
                "consumes": [
                    "application/json"
                ],
                "summary": "update article",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "article id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "article",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateArticleForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/errmsg.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Article"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "delete article by article id",
                "consumes": [
                    "application/json"
                ],
                "summary": "delete article by article id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "article id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/errmsg.Response"
                        }
                    }
                }
            }
        },
        "/api/admin/articles": {
            "get": {
                "description": "list articles",
                "consumes": [
                    "application/json"
                ],
                "summary": "list articles",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "page num",
                        "name": "pageNum",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page size",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/errmsg.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/dto.Pagination"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "list": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/dto.ArticlesListDTO"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ArticleDTO": {
            "type": "object",
            "properties": {
                "allowedComment": {
                    "type": "boolean"
                },
                "categories": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "cover": {
                    "type": "string"
                },
                "editTime": {
                    "type": "string"
                },
                "formatContent": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isTop": {
                    "type": "boolean"
                },
                "likes": {
                    "type": "integer"
                },
                "originalContent": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "private": {
                    "type": "boolean"
                },
                "publishTime": {
                    "type": "string"
                },
                "seoDescription": {
                    "type": "string"
                },
                "seoKeywords": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "slug": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "summary": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "title": {
                    "type": "string"
                },
                "visits": {
                    "type": "integer"
                },
                "wordCount": {
                    "type": "integer"
                }
            }
        },
        "dto.ArticlesListDTO": {
            "type": "object",
            "properties": {
                "allowedComment": {
                    "type": "boolean"
                },
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Category"
                    }
                },
                "commentsNum": {
                    "type": "integer"
                },
                "cover": {
                    "type": "string"
                },
                "editTime": {
                    "type": "string"
                },
                "formatContent": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isTop": {
                    "type": "boolean"
                },
                "likes": {
                    "type": "integer"
                },
                "originalContent": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "private": {
                    "type": "boolean"
                },
                "publishTime": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "summary": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Tag"
                    }
                },
                "title": {
                    "type": "string"
                },
                "visits": {
                    "type": "integer"
                },
                "wordCount": {
                    "type": "integer"
                }
            }
        },
...
        "dto.CreateArticleForm": {
            "type": "object",
            "properties": {
                "allowedComment": {
                    "type": "boolean"
                },
                "categories": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "cover": {
                    "type": "string"
                },
                "formatContent": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isTop": {
                    "type": "boolean"
                },
                "originalContent": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "private": {
                    "type": "boolean"
                },
                "seoDescription": {
                    "type": "string"
                },
                "seoKeywords": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "slug": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "summary": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "title": {
                    "type": "string"
                },
                "wordCount": {
                    "type": "integer"
                }
            }
        },
...
        "dto.Pagination": {
            "type": "object",
            "properties": {
                "current": {
                    "type": "integer"
                },
                "list": {},
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "errmsg.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "msg": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "model.Article": {
            "type": "object",
            "properties": {
                "allowedComment": {
                    "type": "boolean"
                },
                "cover": {
                    "type": "string"
                },
                "editTime": {
                    "type": "string"
                },
                "formatContent": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "likes": {
                    "type": "integer"
                },
                "originalContent": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "private": {
                    "type": "boolean"
                },
                "publishTime": {
                    "type": "string"
                },
                "seoDescription": {
                    "type": "string"
                },
                "seoKeywords": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "summary": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "topPriority": {
                    "type": "integer"
                },
                "visits": {
                    "type": "integer"
                },
                "wordCount": {
                    "type": "integer"
                }
            }
        },
...
        "model.Category": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "model.Tag": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        }
    }
}